spotify: auth token for all requests
Turns out the spotify API now requires an auth token for all
of these requests, not just the ones with the user's private data.

Extract that code into a function and reuse it.

This could probably be designed a little better (e.g., allow
middleware that would transform `key` into `token`), but I'm
just trying to get it working for now.