spotify: change the way it pauses playback
Trying to play a blank playlist causes js errors on the embed
page now.

Instead, open the same song in another tab so that the chrome
extension toggles it to paused. This is pretty hacky, but this
whole chrome extension is pretty fragile since it relies on
spotify's exact implementation at the moment.